fix(): Break the import cycle 1
in the telegram namespace we have a cycle types.ts <-> helpers.ts. We extract the model into separate file so solve this
