+++ src/telegram/actions/checkout.ts
@@ -2,7 +2,8 @@ import { GenericAction } from "./common.js";
 import { DonationStatus } from "../../db/sql/donations.js";
 import { Logger } from "../../logger/index.js";
 import { parseDonationPayload } from "../helpers.js";
-import { type BotMessageModel, TelegramMessagePrefix } from "../types.js";
+import { TelegramMessagePrefix } from "../types.js";
+import type { BotMessageModel } from "../model.js";
 import type { TgCheckoutQuery, TgMessage } from "../api/types.js";
 import type { AnalyticsData } from "../../analytics/ga/types.js";
 import type { PaymentChargeId } from "../api/core.js";
