+++ src/telegram/actions/ignore.ts
@@ -1,7 +1,8 @@
 import { GenericAction } from "./common.js";
 import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
-import { type TelegramMessagePrefix, type BotMessageModel } from "../types.js";
+import type { TelegramMessagePrefix } from "../types.js";
+import type { BotMessageModel } from "../model.js";
 import type { TgMessage } from "../api/types.js";
 
 const logger = new Logger("telegram-bot");
