+++ src/telegram/actions/voice.ts
@@ -9,11 +9,8 @@ import {
 import { collectAnalytics } from "../../analytics/index.js";
 import { TimeMeasure } from "../../common/timer.js";
 import { isBlockedByUser } from "../api/tgerror.js";
-import {
-  type BotMessageModel,
-  type TelegramMessagePrefix,
-  VoiceContentReason,
-} from "../types.js";
+import { type TelegramMessagePrefix, VoiceContentReason } from "../types.js";
+import type { BotMessageModel } from "../model.js";
 import { getFullFileName } from "../../files/index.js";
 import type { TgMessage } from "../api/types.js";
 import type { FileId } from "../api/core.js";
