+++ src/telegram/bot.ts
@@ -7,7 +7,8 @@ import {
 import { Logger } from "../logger/index.js";
 import { type VoiceConverters } from "../recognition/types.js";
 import { TranslationKeys } from "../text/types.js";
-import { BotMessageModel, TelegramMessagePrefix } from "./types.js";
+import { TelegramMessagePrefix } from "./types.js";
+import { BotMessageModel } from "./model.js";
 import { isMessageSupported } from "./helpers.js";
 import { runPromiseWithRetry } from "../common/helpers.js";
 import { getMd5Hash } from "../common/hash.js";
