+++ src/telegram/types.ts
@@ -1,33 +1,10 @@
 import { z } from "zod";
 import { nanoid } from "nanoid";
-import {
-  getChatId,
-  parseDonationPayload,
-  getFullUserName,
-  getGroupName,
-  getRawUserLanguage,
-  getUserLanguage,
-  getUserName,
-  getVoiceDuration,
-  getVoiceFile,
-  isChatGroup,
-  isVideoMessage,
-} from "./helpers.js"; // TODO module cycle with ./types.ts
 import { Logger } from "../logger/index.js";
 import { getTranslator } from "../text/index.js";
-import { type TgMessage } from "./api/types.js";
-import { type AnalyticsData } from "../analytics/ga/types.js";
 import type { LanguageCode } from "../recognition/types.js";
 import type { ValueOf } from "../common/types.js";
-import {
-  type ChatId,
-  type FileId,
-  type MessageId,
-  type MessageThreadId,
-  type PaymentChargeId,
-  TgChatId,
-  type UserId,
-} from "./api/core.js";
+import { type ChatId, TgChatId } from "./api/core.js";
 import type { TgMessageOptions } from "./api/groups/chats/chats-types.js";
 
 export const VoiceContentReason = {
@@ -57,54 +34,6 @@ export const BotCommand = {
 } as const;
 
 export type BotCommandType = ValueOf<typeof BotCommand>;
-
-export class BotMessageModel {
-  public readonly id: MessageId;
-  public readonly chatId: ChatId;
-  public readonly userId?: UserId;
-  public readonly isGroup: boolean;
-  public readonly userName: string;
-  public readonly fullUserName: string;
-  public readonly groupName: string;
-  public readonly voiceFileId?: FileId;
-  public readonly voiceDuration: number;
-  public readonly isVideo: boolean;
-  public readonly userLanguage: LanguageCode;
-  public readonly analytics: AnalyticsData;
-  public readonly donationId: number;
-  public readonly paymentChargeId?: PaymentChargeId;
-  public readonly isSubscriptionPayment: boolean;
-  public readonly forumThreadId?: MessageThreadId;
-
-  constructor(msg: TgMessage, analytics: AnalyticsData) {
-    this.id = msg.message_id;
-    this.chatId = getChatId(msg);
-    this.isGroup = isChatGroup(msg);
-    this.userName = getUserName(msg);
-    this.userId = msg.from?.id;
-    this.fullUserName = getFullUserName(msg);
-    this.groupName = getGroupName(msg);
-    this.voiceFileId = getVoiceFile(msg);
-    this.voiceDuration = getVoiceDuration(msg);
-    this.isVideo = isVideoMessage(msg);
-    this.userLanguage = getUserLanguage(msg);
-    this.donationId = parseDonationPayload(
-      msg.successful_payment?.invoice_payload,
-    ).donationId;
-    this.paymentChargeId = msg.successful_payment?.telegram_payment_charge_id;
-    this.isSubscriptionPayment = msg.successful_payment?.is_recurring ?? false;
-    if (msg.is_topic_message && msg.message_thread_id) {
-      this.forumThreadId = msg.message_thread_id;
-    }
-    this.analytics = analytics
-      .setId(this.chatId)
-      .setLang(getRawUserLanguage(msg));
-  }
-
-  public get name(): string {
-    return this.isGroup ? this.groupName : this.userName;
-  }
-}
 
 export type MessageOptions = {
   lang: LanguageCode;
